chore: clean up files
