--- internals/mocks/fileMock.js
@@ -1,1 +1,0 @@
-export default 'test-file-stub';
